GH-252: Different commands in different files
